feat: implement smith_waterman using a generic function (#6)
* feat: load benchmark data from $FRIZBEE_BENCH_PATH

...instead of from a hardcoding. This will make it easier for different
people to run the benchmarks without modifying the code and recompiling.

The original path is still the default :)

* feat: implement smith_waterman using a generic function

...instead of generating the implementation using a macro.
I've kept the proc macro around for use in benchmarks, but it is not
used in the main lib so it can be deleted safely in the future.

`Bucket` is now a trait, so that the implementation in FixedWidthBucket
can be generic over the string width. This means that Bucket no longer
stores function pointers.

All of the constants were converted to `u8` because `u16` to `u8`
conversion using into is not possible.

* refactor: add some type annotations to make rust-analyzer stop complaining