feat: indicate exact match
This match characteristic is already checked in 'smith_waterman', but refactoring the function to retrieve it from there would likely require significant changes to a mission-critical section, where many optimizations focus on cache locality. Introducing a low-selectivity 8-bit boolean there might not be a wise decision. In the future, if we plan to add more characteristics to 'Match', it might make sense to use a bitmask in 'smith_waterman' instead and then convert it to normal boolean properties in 'Match'.